fixed the numericinput tests folder
