+++ domain/src/test/java/org/oppia/android/domain/classify/rules/numericinput/NumericInputIsLessThanRuleClassifierProviderTest.kt
@@ -1,4 +1,4 @@
-package org.oppia.android.domain.classify.rules
+package org.oppia.android.domain.classify.rules.numericinput
 
 import android.app.Application
 import androidx.test.core.app.ApplicationProvider
@@ -10,7 +10,7 @@ import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.oppia.android.app.model.InteractionObject
-import org.oppia.android.domain.classify.rules.numericinput.NumericInputIsGreaterThanOrEqualToRuleClassifierProvider
+import org.oppia.android.domain.classify.rules.numericinput.NumericInputIsLessThanRuleClassifierProvider
 import org.robolectric.annotation.Config
 import org.robolectric.annotation.LooperMode
 import javax.inject.Inject
@@ -19,11 +19,11 @@ import kotlin.reflect.KClass
 import kotlin.reflect.full.cast
 import kotlin.test.fail
 
-/** Tests for [NumericInputIsGreaterThanOrEqualToRuleClassifierProvider]. */
+/** Tests for [NumericInputIsLessThanRuleClassifierProvider]. */
 @RunWith(AndroidJUnit4::class)
 @LooperMode(LooperMode.Mode.PAUSED)
 @Config(manifest = Config.NONE)
-class NumericInputIsGreaterThanOrEqualToRuleClassifierProviderTest {
+class NumericInputIsLessThanRuleClassifierProviderTest {
 
   private val POSITIVE_REAL_VALUE_1_5 = createReal(value = 1.5)
   private val POSITIVE_REAL_VALUE_3_5 = createReal(value = 3.5)
@@ -36,11 +36,11 @@ class NumericInputIsGreaterThanOrEqualToRuleClassifierProviderTest {
   private val NEGATIVE_INT_VALUE_3 = createInt(value = -3)
 
   @Inject
-  internal lateinit var numericInputIsGreaterThanOrEqualToRuleClassifierProvider:
-    NumericInputIsGreaterThanOrEqualToRuleClassifierProvider
+  internal lateinit var numericInputIsLessThanRuleClassifierProvider:
+    NumericInputIsLessThanRuleClassifierProvider
 
-  private val inputIsGreaterThanOrEqualToRuleClassifier by lazy {
-    numericInputIsGreaterThanOrEqualToRuleClassifierProvider.createRuleClassifier()
+  private val inputIsLessThanRuleClassifier by lazy {
+    numericInputIsLessThanRuleClassifierProvider.createRuleClassifier()
   }
 
   @Before
@@ -49,113 +49,113 @@ class NumericInputIsGreaterThanOrEqualToRuleClassifierProviderTest {
   }
 
   @Test
-  fun testPositiveRealAnswer_positiveRealInput_sameExactValues_answerGreaterOrEqual() {
+  fun testPositiveRealAnswer_positiveRealInput_sameExactValues_answerNotLesser() {
     val inputs = mapOf("x" to POSITIVE_REAL_VALUE_1_5)
 
     val matches =
-      inputIsGreaterThanOrEqualToRuleClassifier
+      inputIsLessThanRuleClassifier
         .matches(answer = POSITIVE_REAL_VALUE_1_5, inputs = inputs)
 
-    assertThat(matches).isTrue()
+    assertThat(matches).isFalse()
   }
 
   @Test
-  fun testNegativeRealAnswer_negativeRealInput_sameExactValues_answerGreaterOrEqual() {
+  fun testNegativeRealAnswer_negativeRealInput_sameExactValues_answerNotLesser() {
     val inputs = mapOf("x" to NEGATIVE_REAL_VALUE_1_5)
 
     val matches =
-      inputIsGreaterThanOrEqualToRuleClassifier
+      inputIsLessThanRuleClassifier
         .matches(answer = NEGATIVE_REAL_VALUE_1_5, inputs = inputs)
 
-    assertThat(matches).isTrue()
+    assertThat(matches).isFalse()
   }
 
   @Test
-  fun testPositiveRealAnswer_positiveRealInput_answerValueGreater_answerGreaterOrEqual() {
-    val inputs = mapOf("x" to POSITIVE_REAL_VALUE_1_5)
-
-    val matches =
-      inputIsGreaterThanOrEqualToRuleClassifier
-        .matches(answer = POSITIVE_REAL_VALUE_3_5, inputs = inputs)
-
-    assertThat(matches).isTrue()
-  }
-
-  @Test
-  fun testPositiveRealAnswer_positiveRealInput_answerValueSmaller_answerNotGreaterOrEqual() {
+  fun testPositiveRealAnswer_positiveRealInput_answerValueLesser_answerLesser() {
     val inputs = mapOf("x" to POSITIVE_REAL_VALUE_3_5)
 
     val matches =
-      inputIsGreaterThanOrEqualToRuleClassifier
+      inputIsLessThanRuleClassifier
         .matches(answer = POSITIVE_REAL_VALUE_1_5, inputs = inputs)
-
-    assertThat(matches).isFalse()
-  }
-
-  @Test
-  fun testNegativeRealAnswer_negativeRealInput_answerValueGreater_answerGreaterOrEqual() {
-    val inputs = mapOf("x" to NEGATIVE_REAL_VALUE_3_5)
-
-    val matches =
-      inputIsGreaterThanOrEqualToRuleClassifier
-        .matches(answer = NEGATIVE_REAL_VALUE_1_5, inputs = inputs)
 
     assertThat(matches).isTrue()
   }
 
   @Test
-  fun testNegativeRealAnswer_negativeRealInput_answerValueSmaller_answerNotGreaterOrEqual() {
-    val inputs = mapOf("x" to NEGATIVE_REAL_VALUE_1_5)
-
-    val matches =
-      inputIsGreaterThanOrEqualToRuleClassifier
-        .matches(answer = NEGATIVE_REAL_VALUE_3_5, inputs = inputs)
-
-    assertThat(matches).isFalse()
-  }
-
-  @Test
-  fun testNegativeRealAnswer_positiveRealInput_answerValueSmaller_answerNotGreaterOrEqual() {
+  fun testPositiveRealAnswer_positiveRealInput_answerValueGreater_answerNotLesser() {
     val inputs = mapOf("x" to POSITIVE_REAL_VALUE_1_5)
 
     val matches =
-      inputIsGreaterThanOrEqualToRuleClassifier
-        .matches(answer = NEGATIVE_REAL_VALUE_3_5, inputs = inputs)
+      inputIsLessThanRuleClassifier
+        .matches(answer = POSITIVE_REAL_VALUE_3_5, inputs = inputs)
 
     assertThat(matches).isFalse()
   }
 
   @Test
-  fun testPositiveRealAnswer_negativeRealInput_answerValueGreater_answerGreaterOrEqual() {
+  fun testNegativeRealAnswer_negativeRealInput_answerValueLesser_answerLesser() {
     val inputs = mapOf("x" to NEGATIVE_REAL_VALUE_1_5)
 
     val matches =
-      inputIsGreaterThanOrEqualToRuleClassifier
-        .matches(answer = POSITIVE_REAL_VALUE_1_5, inputs = inputs)
+      inputIsLessThanRuleClassifier
+        .matches(answer = NEGATIVE_REAL_VALUE_3_5, inputs = inputs)
 
     assertThat(matches).isTrue()
   }
 
   @Test
-  fun testPositiveIntAnswer_negativeIntInput_answerValueGreater_answerGreaterOrEqual() {
+  fun testNegativeRealAnswer_negativeRealInput_answerValueGreater_answerNotLesser() {
+    val inputs = mapOf("x" to NEGATIVE_REAL_VALUE_3_5)
+
+    val matches =
+      inputIsLessThanRuleClassifier
+        .matches(answer = NEGATIVE_REAL_VALUE_1_5, inputs = inputs)
+
+    assertThat(matches).isFalse()
+  }
+
+  @Test
+  fun testNegativeRealAnswer_positiveRealInput_answerValueLesser_answerLesser() {
+    val inputs = mapOf("x" to POSITIVE_REAL_VALUE_1_5)
+
+    val matches =
+      inputIsLessThanRuleClassifier
+        .matches(answer = NEGATIVE_REAL_VALUE_3_5, inputs = inputs)
+
+    assertThat(matches).isTrue()
+  }
+
+  @Test
+  fun testPositiveRealAnswer_negativeRealInput_answerValueGreater_answerNotLesser() {
+    val inputs = mapOf("x" to NEGATIVE_REAL_VALUE_1_5)
+
+    val matches =
+      inputIsLessThanRuleClassifier
+        .matches(answer = POSITIVE_REAL_VALUE_1_5, inputs = inputs)
+
+    assertThat(matches).isFalse()
+  }
+
+  @Test
+  fun testPositiveIntAnswer_negativeIntInput_answerValueGreater_answerNotLesser() {
     val inputs = mapOf("x" to NEGATIVE_INT_VALUE_3)
 
     val matches =
-      inputIsGreaterThanOrEqualToRuleClassifier
+      inputIsLessThanRuleClassifier
         .matches(answer = POSITIVE_INT_VALUE_1, inputs = inputs)
 
-    assertThat(matches).isTrue()
+    assertThat(matches).isFalse()
   }
 
   @Test
-  fun testNegativeIntAnswer_positiveIntInput_answerValueSmaller_answerNotGreaterOrEqual() {
+  fun testNegativeIntAnswer_positiveIntInput_answerValueLesser_answerLesser() {
     val inputs = mapOf("x" to POSITIVE_INT_VALUE_3)
 
     val matches =
-      inputIsGreaterThanOrEqualToRuleClassifier
+      inputIsLessThanRuleClassifier
         .matches(answer = NEGATIVE_INT_VALUE_1, inputs = inputs)
 
-    assertThat(matches).isFalse()
+    assertThat(matches).isTrue()
   }
 
   @Test
@@ -163,7 +163,7 @@ class NumericInputIsGreaterThanOrEqualToRuleClassifierProviderTest {
     val inputs = mapOf("y" to POSITIVE_REAL_VALUE_1_5)
 
     val exception = assertThrows(IllegalStateException::class) {
-      inputIsGreaterThanOrEqualToRuleClassifier
+      inputIsLessThanRuleClassifier
         .matches(answer = POSITIVE_REAL_VALUE_1_5, inputs = inputs)
     }
 
@@ -177,8 +177,36 @@ class NumericInputIsGreaterThanOrEqualToRuleClassifierProviderTest {
     val inputs = mapOf("x" to STRING_VALUE)
 
     val exception = assertThrows(IllegalStateException::class) {
-      inputIsGreaterThanOrEqualToRuleClassifier
+      inputIsLessThanRuleClassifier
         .matches(answer = POSITIVE_REAL_VALUE_1_5, inputs = inputs)
+    }
+
+    assertThat(exception)
+      .hasMessageThat()
+      .contains("Expected input value to be of type REAL not NORMALIZED_STRING")
+  }
+
+  @Test
+  fun testIntAnswer_missingInput_throwsException() {
+    val inputs = mapOf("y" to POSITIVE_INT_VALUE_1)
+
+    val exception = assertThrows(IllegalStateException::class) {
+      inputIsLessThanRuleClassifier
+        .matches(answer = POSITIVE_INT_VALUE_3, inputs = inputs)
+    }
+
+    assertThat(exception)
+      .hasMessageThat()
+      .contains("Expected classifier inputs to contain parameter with name 'x'")
+  }
+
+  @Test
+  fun testIntAnswer_stringInput_throwsException() {
+    val inputs = mapOf("x" to STRING_VALUE)
+
+    val exception = assertThrows(IllegalStateException::class) {
+      inputIsLessThanRuleClassifier
+        .matches(answer = NEGATIVE_INT_VALUE_1, inputs = inputs)
     }
 
     assertThat(exception)
@@ -199,7 +227,7 @@ class NumericInputIsGreaterThanOrEqualToRuleClassifierProviderTest {
   }
 
   private fun setUpTestApplicationComponent() {
-    DaggerNumericInputIsGreaterThanOrEqualToRuleClassifierProviderTest_TestApplicationComponent
+    DaggerNumericInputIsLessThanRuleClassifierProviderTest_TestApplicationComponent
       .builder()
       .setApplication(ApplicationProvider.getApplicationContext())
       .build()
@@ -232,6 +260,6 @@ class NumericInputIsGreaterThanOrEqualToRuleClassifierProviderTest {
       fun build(): TestApplicationComponent
     }
 
-    fun inject(test: NumericInputIsGreaterThanOrEqualToRuleClassifierProviderTest)
+    fun inject(test: NumericInputIsLessThanRuleClassifierProviderTest)
   }
 }
